FIR: migrate type-related utils
