+++ compiler/fir/resolve/src/org/jetbrains/kotlin/fir/resolve/transformers/body/resolve/FirDeclarationsResolveTransformer.kt
@@ -6,11 +6,6 @@
 package org.jetbrains.kotlin.fir.resolve.transformers.body.resolve
 
 import org.jetbrains.kotlin.descriptors.ClassKind
-import org.jetbrains.kotlin.descriptors.Visibilities.Internal
-import org.jetbrains.kotlin.descriptors.Visibilities.Private
-import org.jetbrains.kotlin.descriptors.Visibilities.Protected
-import org.jetbrains.kotlin.descriptors.Visibilities.Public
-import org.jetbrains.kotlin.descriptors.Visibility
 import org.jetbrains.kotlin.fir.*
 import org.jetbrains.kotlin.fir.declarations.*
 import org.jetbrains.kotlin.fir.declarations.builder.buildValueParameter
@@ -34,9 +29,7 @@ import org.jetbrains.kotlin.fir.resolve.substitution.ConeSubstitutor
 import org.jetbrains.kotlin.fir.resolve.transformers.*
 import org.jetbrains.kotlin.fir.scopes.impl.FirLocalScope
 import org.jetbrains.kotlin.fir.scopes.impl.FirMemberTypeParameterScope
-import org.jetbrains.kotlin.fir.scopes.impl.withReplacedConeType
 import org.jetbrains.kotlin.fir.symbols.constructStarProjectedType
-import org.jetbrains.kotlin.fir.symbols.impl.ConeClassLookupTagWithFixedSymbol
 import org.jetbrains.kotlin.fir.symbols.impl.FirVariableSymbol
 import org.jetbrains.kotlin.fir.types.*
 import org.jetbrains.kotlin.fir.types.builder.buildErrorTypeRef
@@ -44,7 +37,6 @@ import org.jetbrains.kotlin.fir.types.builder.buildImplicitTypeRef
 import org.jetbrains.kotlin.fir.types.builder.buildResolvedTypeRef
 import org.jetbrains.kotlin.fir.visitors.*
 import org.jetbrains.kotlin.name.Name
-import org.jetbrains.kotlin.types.TypeApproximatorConfiguration
 import org.jetbrains.kotlin.utils.addToStdlib.safeAs
 
 open class FirDeclarationsResolveTransformer(transformer: FirBodyResolveTransformer) : FirPartialBodyResolveTransformer(transformer) {
@@ -511,7 +503,9 @@ open class FirDeclarationsResolveTransformer(transformer: FirBodyResolveTransfor
                 result.transformReturnTypeRef(
                     transformer,
                     withExpectedType(
-                        body.resultType.approximateTypeIfNeeded(simpleFunction?.visibility, simpleFunction?.isInline == true)
+                        body.resultType.approximateTypeIfNeeded(
+                            inferenceComponents.approximator, simpleFunction?.visibility, simpleFunction?.isInline == true
+                        )
                     )
                 )
             } else {
@@ -923,7 +917,9 @@ open class FirDeclarationsResolveTransformer(transformer: FirBodyResolveTransfor
                     }
                     variable.transformReturnTypeRef(
                         transformer,
-                        withExpectedType(expectedType.approximateTypeIfNeeded((variable as? FirProperty)?.visibility))
+                        withExpectedType(
+                            expectedType.approximateTypeIfNeeded(inferenceComponents.approximator, (variable as? FirProperty)?.visibility)
+                        )
                     )
                 }
                 variable.getter != null && variable.getter !is FirDefaultPropertyAccessor -> {
@@ -945,7 +941,9 @@ open class FirDeclarationsResolveTransformer(transformer: FirBodyResolveTransfor
                     }
                     variable.transformReturnTypeRef(
                         transformer,
-                        withExpectedType(expectedType?.approximateTypeIfNeeded((variable as? FirProperty)?.visibility))
+                        withExpectedType(
+                            expectedType?.approximateTypeIfNeeded(inferenceComponents.approximator, (variable as? FirProperty)?.visibility)
+                        )
                     )
                 }
                 else -> {
@@ -968,86 +966,6 @@ open class FirDeclarationsResolveTransformer(transformer: FirBodyResolveTransfor
         }
     }
 
-    private fun ConeKotlinType.requiresApproximationInPublicPosition(): Boolean {
-        return when (this) {
-            is ConeIntegerLiteralType,
-            is ConeCapturedType,
-            is ConeDefinitelyNotNullType,
-            is ConeIntersectionType -> true
-            is ConeClassLikeType -> typeArguments.any {
-                it is ConeKotlinTypeProjection && it.type.requiresApproximationInPublicPosition()
-            }
-            else -> false
-        }
-    }
-
-    private fun FirTypeRef.approximateTypeIfNeeded(
-        containingCallableVisibility: Visibility?,
-        isInlineFunction: Boolean = false
-    ): FirTypeRef {
-        val approximatedType = if (this is FirResolvedTypeRef &&
-            (containingCallableVisibility == Public || containingCallableVisibility == Protected)
-        ) {
-            if (type.requiresApproximationInPublicPosition()) {
-                this.withReplacedConeType(
-                    inferenceComponents.approximator.approximateToSuperType(
-                        this.type, TypeApproximatorConfiguration.PublicDeclaration
-                    ) as? ConeKotlinType
-                )
-            } else {
-                this
-            }
-        } else {
-            this
-        }
-        return approximatedType.hideLocalTypeIfNeeded(containingCallableVisibility, isInlineFunction)
-    }
-
-    /*
-     * Suppose a function without an explicit return type just returns an anonymous object:
-     *
-     *   fun foo(...) = object : ObjectSuperType {
-     *     override fun ...
-     *   }
-     *
-     * Without unwrapping, the return type ended up with that anonymous object (<no name provided>), while the resolved super type, which
-     * acts like an implementing interface, is a better fit. In fact, exposing an anonymous object types is prohibited for certain cases,
-     * e.g., KT-33917. We can also apply this to any local types.
-     */
-    private fun FirTypeRef.hideLocalTypeIfNeeded(
-        containingCallableVisibility: Visibility?,
-        isInlineFunction: Boolean = false
-    ): FirTypeRef {
-        if (containingCallableVisibility == null) {
-            return this
-        }
-        // Approximate types for non-private (all but package private or private) members.
-        // Also private inline functions, as per KT-33917.
-        if (containingCallableVisibility == Public ||
-            containingCallableVisibility == Protected ||
-            containingCallableVisibility == Internal ||
-            (containingCallableVisibility == Private && isInlineFunction)
-        ) {
-            val firClass =
-                (((this as? FirResolvedTypeRef)
-                    ?.type as? ConeClassLikeType)
-                    ?.lookupTag as? ConeClassLookupTagWithFixedSymbol)
-                    ?.symbol?.fir
-            if (firClass?.classId?.isLocal != true) {
-                return this
-            }
-            if (firClass.superTypeRefs.size > 1) {
-                return buildErrorTypeRef {
-                    diagnostic = ConeSimpleDiagnostic("Cannot hide local type ${firClass.render()}")
-                }
-            }
-            val superType = firClass.superTypeRefs.single()
-            if (superType is FirResolvedTypeRef && !superType.isAny) {
-                return superType
-            }
-        }
-        return this
-    }
 
     private object ImplicitToErrorTypeTransformer : FirTransformer<Nothing?>() {
         override fun <E : FirElement> transformElement(element: E, data: Nothing?): CompositeTransformResult<E> {
