+++ compiler/fir/resolve/src/org/jetbrains/kotlin/fir/types/TypeUtils.kt
@@ -5,12 +5,27 @@
 
 package org.jetbrains.kotlin.fir.types
 
+import org.jetbrains.kotlin.descriptors.Visibilities.Internal
+import org.jetbrains.kotlin.descriptors.Visibilities.Private
+import org.jetbrains.kotlin.descriptors.Visibilities.Protected
+import org.jetbrains.kotlin.descriptors.Visibilities.Public
+import org.jetbrains.kotlin.descriptors.Visibility
+import org.jetbrains.kotlin.fir.FirFakeSourceElementKind
 import org.jetbrains.kotlin.fir.FirSession
+import org.jetbrains.kotlin.fir.declarations.classId
+import org.jetbrains.kotlin.fir.diagnostics.ConeSimpleDiagnostic
+import org.jetbrains.kotlin.fir.fakeElement
+import org.jetbrains.kotlin.fir.render
 import org.jetbrains.kotlin.fir.resolve.fullyExpandedType
+import org.jetbrains.kotlin.fir.symbols.impl.ConeClassLookupTagWithFixedSymbol
+import org.jetbrains.kotlin.fir.types.builder.buildErrorTypeRef
+import org.jetbrains.kotlin.fir.types.builder.buildResolvedTypeRef
 import org.jetbrains.kotlin.fir.types.impl.ConeClassLikeTypeImpl
 import org.jetbrains.kotlin.fir.types.impl.ConeTypeParameterTypeImpl
 import org.jetbrains.kotlin.resolve.calls.NewCommonSuperTypeCalculator
 import org.jetbrains.kotlin.types.AbstractStrictEqualityTypeChecker
+import org.jetbrains.kotlin.types.AbstractTypeApproximator
+import org.jetbrains.kotlin.types.TypeApproximatorConfiguration
 
 fun ConeInferenceContext.commonSuperTypeOrNull(types: List<ConeKotlinType>): ConeKotlinType? {
     return when (types.size) {
@@ -171,3 +186,114 @@ fun FirTypeRef.isUnsafeVarianceType(session: FirSession): Boolean {
     return coneTypeSafe<ConeKotlinType>()?.isUnsafeVarianceType(session) == true
 }
 
+// Unlike other cases, return types may be implicit, i.e. unresolved
+// But in that cases newType should also be `null`
+fun FirTypeRef.withReplacedReturnType(newType: ConeKotlinType?): FirTypeRef {
+    require(this is FirResolvedTypeRef || newType == null)
+    if (newType == null) return this
+
+    return buildResolvedTypeRef {
+        source = this@withReplacedReturnType.source
+        type = newType
+        annotations += this@withReplacedReturnType.annotations
+    }
+}
+
+fun FirTypeRef.withReplacedConeType(
+    newType: ConeKotlinType?,
+    firFakeSourceElementKind: FirFakeSourceElementKind? = null
+): FirResolvedTypeRef {
+    require(this is FirResolvedTypeRef)
+    if (newType == null) return this
+
+    return buildResolvedTypeRef {
+        source = if (firFakeSourceElementKind != null)
+            this@withReplacedConeType.source?.fakeElement(firFakeSourceElementKind)
+        else
+            this@withReplacedConeType.source
+        type = newType
+        annotations += this@withReplacedConeType.annotations
+    }
+}
+
+private fun ConeKotlinType.requiresApproximationInPublicPosition(): Boolean {
+    return when (this) {
+        is ConeIntegerLiteralType,
+        is ConeCapturedType,
+        is ConeDefinitelyNotNullType,
+        is ConeIntersectionType -> true
+        is ConeClassLikeType -> typeArguments.any {
+            it is ConeKotlinTypeProjection && it.type.requiresApproximationInPublicPosition()
+        }
+        else -> false
+    }
+}
+
+fun FirTypeRef.approximateTypeIfNeeded(
+    approximator: AbstractTypeApproximator,
+    containingCallableVisibility: Visibility?,
+    isInlineFunction: Boolean = false
+): FirTypeRef {
+    val approximatedType = if (this is FirResolvedTypeRef &&
+        (containingCallableVisibility == Public || containingCallableVisibility == Protected)
+    ) {
+        if (type.requiresApproximationInPublicPosition()) {
+            this.withReplacedConeType(
+                approximator.approximateToSuperType(
+                    this.type, TypeApproximatorConfiguration.PublicDeclaration
+                ) as? ConeKotlinType
+            )
+        } else {
+            this
+        }
+    } else {
+        this
+    }
+    return approximatedType.hideLocalTypeIfNeeded(containingCallableVisibility, isInlineFunction)
+}
+
+/*
+ * Suppose a function without an explicit return type just returns an anonymous object:
+ *
+ *   fun foo(...) = object : ObjectSuperType {
+ *     override fun ...
+ *   }
+ *
+ * Without unwrapping, the return type ended up with that anonymous object (<no name provided>), while the resolved super type, which
+ * acts like an implementing interface, is a better fit. In fact, exposing an anonymous object types is prohibited for certain cases,
+ * e.g., KT-33917. We can also apply this to any local types.
+ */
+private fun FirTypeRef.hideLocalTypeIfNeeded(
+    containingCallableVisibility: Visibility?,
+    isInlineFunction: Boolean = false
+): FirTypeRef {
+    if (containingCallableVisibility == null) {
+        return this
+    }
+    // Approximate types for non-private (all but package private or private) members.
+    // Also private inline functions, as per KT-33917.
+    if (containingCallableVisibility == Public ||
+        containingCallableVisibility == Protected ||
+        containingCallableVisibility == Internal ||
+        (containingCallableVisibility == Private && isInlineFunction)
+    ) {
+        val firClass =
+            (((this as? FirResolvedTypeRef)
+                ?.type as? ConeClassLikeType)
+                ?.lookupTag as? ConeClassLookupTagWithFixedSymbol)
+                ?.symbol?.fir
+        if (firClass?.classId?.isLocal != true) {
+            return this
+        }
+        if (firClass.superTypeRefs.size > 1) {
+            return buildErrorTypeRef {
+                diagnostic = ConeSimpleDiagnostic("Cannot hide local type ${firClass.render()}")
+            }
+        }
+        val superType = firClass.superTypeRefs.single()
+        if (superType is FirResolvedTypeRef && !superType.isAny) {
+            return superType
+        }
+    }
+    return this
+}
